[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 2220 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -13,7 +13,7 @@ int main(){
     }
     while(low <= high){
         ll mid = (low+high)/2;
-        int total = 0;
+        ll total = 0;
         for(auto l : line){
             total += l/mid;
         }
